Use fakeOrg data from TaskCluster Secrets instead of static file
