+++ .neutrinorc.js
@@ -16,7 +16,6 @@ module.exports = {
     [
       '@neutrinojs/copy', {
         patterns: [
-          { from: 'src/static/fakeOrg.json', to: 'people.json' },
           { from: 'src/static/triageOwners.json', to: 'triageOwners.json' },
         ],
       },
