+++ src/utils/getAllReportees.js
@@ -1,4 +1,40 @@
-import getOrgChart from './getOrgChart';
+import config from '../config';
+
+const buildOrgChartData = (people) => {
+  const org = {};
+  people.forEach((person) => {
+    const { mail } = person;
+    if (!org[mail]) {
+      org[mail] = person;
+      org[mail].reportees = [];
+    } else {
+      org[mail] = {
+        ...person,
+        reportees: org[mail].reportees,
+      };
+    }
+    const { manager } = person;
+    if (manager) {
+      const managerLDAPemail = manager.dn.split('mail=')[1].split(',o=')[0];
+      if (org[managerLDAPemail]) {
+        org[managerLDAPemail].reportees.push(mail);
+      } else {
+        org[managerLDAPemail] = {
+          reportees: [mail],
+        };
+      }
+    }
+    if (!org[mail].bugzillaEmail) {
+      org[mail].bugzillaEmail = mail;
+    }
+  });
+  return org;
+};
+
+const getOrgChart = async (secretsClient) => {
+  const { secret } = await await secretsClient.get(config.taskclusterSecrets.fakeOrg);
+  return buildOrgChartData(secret.employees);
+};
 
 const findReportees = (completeOrg, ldapEmail) => {
   let allReportees = {};
@@ -13,8 +49,8 @@ const findReportees = (completeOrg, ldapEmail) => {
   return allReportees;
 };
 
-const getAllReportees = async (ldapEmail) => {
-  const completeOrg = await getOrgChart();
+const getAllReportees = async (secretsClient, ldapEmail) => {
+  const completeOrg = await getOrgChart(secretsClient);
   return findReportees(completeOrg, ldapEmail);
 };
