--- src/utils/getOrgChart.js
@@ -1,37 +1,0 @@
-const orgChart = (people) => {
-  const org = {};
-  people.forEach((person) => {
-    const { mail } = person;
-    if (!org[mail]) {
-      org[mail] = person;
-      org[mail].reportees = [];
-    } else {
-      org[mail] = {
-        ...person,
-        reportees: org[mail].reportees,
-      };
-    }
-    const { manager } = person;
-    if (manager) {
-      const managerLDAPemail = manager.dn.split('mail=')[1].split(',o=')[0];
-      if (org[managerLDAPemail]) {
-        org[managerLDAPemail].reportees.push(mail);
-      } else {
-        org[managerLDAPemail] = {
-          reportees: [mail],
-        };
-      }
-    }
-    if (!org[mail].bugzillaEmail) {
-      org[mail].bugzillaEmail = mail;
-    }
-  });
-  return org;
-};
-
-const getOrgChart = async () => {
-  const people = await (await fetch('people.json')).json();
-  return orgChart(people);
-};
-
-export default getOrgChart;
